+++ src/main/java/com/bitbank/responses/v1/MessageResponse.java
@@ -1,0 +1,35 @@
+package com.bitbank.responses.v1;
+
+/**
+ * Generic class for ResponseEntity
+ * 
+ * @author sineverba
+ *
+ */
+public class MessageResponse {
+
+	private String status;
+	private String message;
+
+	public MessageResponse(String status, String message) {
+		this.setStatus(status);
+		this.setMessage(message);
+	}
+
+	public String getStatus() {
+		return status;
+	}
+
+	private void setStatus(String status) {
+		this.status = status;
+	}
+
+	public String getMessage() {
+		return message;
+	}
+
+	private void setMessage(String message) {
+		this.message = message;
+	}
+
+}