+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 + Refactor docker-compose for production
 + Refactor Response Entity for PingController
 + Migrate to custom Sonarqube. Fix double CSRF disable
++ Refactor Message Entity for Registration Controller
 
 ## 0.5.0
 + Enable registration URL only when APP properties is true